requirements.txt
fastapi
uvicorn
pydantic
pandas
numpy
langchain
ipykernel # this required only in local to run ipynote notebook
#load_dotenv
python-dotenv
langchain-openai
langchain-core
bs4 #beautiful soup
langchain_community
faiss-cpu
pypdf
arxiv
pymupdf
wikipedia
lxml
langchain_huggingface
langchain-groq
langgraph
langgraph_cli[inmem]
#pygraphviz
tavily-python
duckduckgo_search
pymilvus[lite]
milvus-lite
chromadb
langchain-anthropic
langchain-google-genai 
google-generativeai
langchain-ollama
streamlit
langchain_tavily
huggingface_hub[hf_xet]
#neo4j
langchain-neo4j 
langchain_experimental
langchain_deepseek
mcp[cli]